Update semua file yang berubah

+++ day58.cpp
@@ -9,7 +9,7 @@ struct Transaksi
 {
     string tanggal;
     string Keterangan;
-    double jumlah;
+    long double jumlah;
     bool pemasukan; 
 };
